% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsos.R
\name{lsos}
\alias{lsos}
\title{Memory manage trick courtesy of Stack Exchange}
\usage{
lsos(..., n = 10)
}
\arguments{
\item{n}{How many objects to show.}
}
\description{
Memory manage trick courtesy of Stack Exchange
}
\examples{
\dontrun{
lsos(n = 10)
}

}
