% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollDice.R
\name{rollDice}
\alias{rollDice}
\title{A dice roller for DnD}
\usage{
rollDice(dice, modifier = NULL)
}
\arguments{
\item{dice}{A vector of dice values.}

\item{modifier}{Optional modifier.}
}
\description{
A dice roller for DnD
}
\examples{
# Roll 2d10 + 7
rollDice(rep(10,2), 7)

}
