% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitBill.R
\name{splitBill}
\alias{splitBill}
\title{Takes orders as a vector, total price after tax/delivery/tip}
\usage{
splitBill(orders, total, names = NULL)
}
\arguments{
\item{orders}{Vector of orders as numerics, sums are allowed.}

\item{total}{Total for the order, after all fees.}

\item{names}{Optional vector of names for pretty output}
}
\description{
Takes orders as a vector, total price after tax/delivery/tip
}
\examples{
splitBill(c(3.30, 1.90, 3.80, 8.25 + 1.90, 10.25),
35.36, names = c("Malcolm", "Russell", "Yandi", "Johnny", "Bryan"))

}
