% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateR.R
\name{updateR}
\alias{updateR}
\title{A function to update R without losing packages, intended for Macs.}
\usage{
updateR()
}
\description{
A function to update R without losing packages, intended for Macs.
}
\examples{
updateR()

}
